requirements.txt
fastapi[all]
uvicorn[standard]
sqlalchemy
asyncpg
alembic
python-jose[cryptography]
passlib[argon2]
python-multipart
pydantic[email]
pydantic-settings
greenlet
